* fixed ProcessRunner.CleanupTmpDir to restore ProcessStartInfo TMP value in case it's re-used (like MSFraggerSearchEngine does)

+++ pwiz_tools/Shared/CommonUtil/SystemUtil/ProcessRunner.cs
@@ -167,7 +167,7 @@ namespace pwiz.Common.SystemUtil
                         {
                             proc.Kill();
                             progress.UpdateProgress(status = status.Cancel());
-                            CleanupTmpDir(); // Clean out any tempfiles left behind, if forceTempfilesCleanup was set
+                            CleanupTmpDir(psi); // Clean out any tempfiles left behind, if forceTempfilesCleanup was set
                             return;
                         }
 
@@ -249,7 +249,7 @@ namespace pwiz.Common.SystemUtil
                 if (!proc.HasExited)
                     try { proc.Kill(); } catch (InvalidOperationException) { }
 
-                CleanupTmpDir(); // Clean out any tempfiles left behind, if forceTempfilesCleanup was set
+                CleanupTmpDir(psi); // Clean out any tempfiles left behind, if forceTempfilesCleanup was set
             }
         }
 
@@ -264,7 +264,7 @@ namespace pwiz.Common.SystemUtil
         }
 
         // Clean out any tempfiles left behind, if forceTempfilesCleanup was set
-        private void CleanupTmpDir()
+        private void CleanupTmpDir(ProcessStartInfo psi)
         {
             if (!string.IsNullOrEmpty(_tmpDirForCleanup))
             {
@@ -273,7 +273,9 @@ namespace pwiz.Common.SystemUtil
                 {
                     try
                     {
-                        Directory.Delete(_tmpDirForCleanup, true);
+                        if (Directory.Exists(_tmpDirForCleanup))
+                            Directory.Delete(_tmpDirForCleanup, true);
+                        psi.Environment[@"TMP"] = Path.GetDirectoryName(_tmpDirForCleanup); // restore previous TMP value in case ProcessStartInfo is re-used
                         return;
                     }
                     catch (Exception e)
